Initial commit of the Projection Iterator.
This header provides an iterator adaptor that applies a projection to a
random access iterator.  In principle, it can be extended to bidirectional
iterators, but this version doesn't take that step.

The demo program that comes with it demonstrates sorting a vector such that
the first half of the sorted elements end up on even indices in increasing
order and the second half of sorted elements end up on odd indices in
decreasing order.  The sort just requires std::sort with no post-copy or
post-permute.