Add LICENSE, README.md, and license info
Added licensing info to the source files, as well as a LICENSE file
and a README.md that describes what all this is.

Note: There is no Makefile or project or anything like that.  Just
compile the `.cc` file in such a way that it can see the `.hh` file.