Take Callable by value, not reference
I realized taking Callable& was a bad idea, as someone might provide a
lambda as a projection.  That would immediately run into lifetime issues.
So, now I just take it by copy.

If this were performance sensitive in any way, I might think about this
harder.  For this toy, I'm not too worried.  Obviously, if someone wants
to turn this into a standard library component, it needs more thought to
edge cases.